requirements.txt
scrapy-proxies==0.3
